ui/jupyter-notebook/requirements.txt
# Jupyter Notebook MCP Client Dependencies

# Core dependencies
anthropic>=0.40.0
python-dotenv>=1.0.0

# Jupyter
jupyter>=1.0.0
ipykernel>=6.0.0
notebook>=7.0.0

# Data analysis
pandas>=2.0.0
numpy>=1.23.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.0.0

# Optional: For enhanced visualizations
ipywidgets>=8.0.0
